requirements.txt
Django=3.0.7
ldap3=2.6.1
python-decouple=3.3